requirements.txt
unstructured[all-docs]==0.10.15
tiktoken
requests
datetime
streamlit
python-dotenv
pathway
llm_app